+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Improved
 - Changed get of delivery countries from API instead of code
+- Added option to select shipping method title type on Checkout page
 
 ### Updated
 - API library to v1.0.3